feat- refactor the file structure for components of layout related components
